refactor annotate_polygon.py
