+++ masks/case035/timepoint000/mask.json
@@ -1,0 +1,1 @@
+{"polygons": [{"vertices": [[1124, 856], [1424, 933], [1232, 1109], [1011, 1042], [1034, 895]]}]}
