+++ masks/case035/timepoint001/mask.json
@@ -1,0 +1,1 @@
+{"polygons": [{"vertices": [[1088, 1039], [1399, 1122], [1184, 1321]]}]}
