+++ src/polygon_annotator/constants.py
@@ -1,0 +1,14 @@
+LEFT_KEYCODE = 80
+RIGHT_KEYCODE = 79
+R_KEYCODE = 21
+U_KEYCODE = 24
+ENTER_KEYCODE = 40
+
+HELP_TEXT = (
+    "[b]Controls[/b]\n"
+    "- Left click        : add polygon vertex (RIGHT panel)\n"
+    "- Right click/Enter : finalize polygon + fill (>= 3 vertices)\n"
+    "- u                 : undo last vertex\n"
+    "- r                 : delete saved mask + clear overlays\n"
+    "- Left / Right      : prev / next sample (auto-save)\n"
+)
